8357086: Addressed reviewer's comments

+++ src/hotspot/share/jfr/jni/jfrJniMethod.cpp
@@ -412,7 +412,7 @@ JVM_ENTRY_NO_ENV(jlong, jfr_host_total_memory(JNIEnv* env, jclass jvm))
 #ifdef LINUX
   // We want the host memory, not the container limit.
   // os::physical_memory() would return the container limit.
-  return os::Linux::physical_memory();
+  return static_cast<jlong>(os::Linux::physical_memory());
 #else
   return static_cast<jlong>(os::physical_memory());
 #endif
